Refactored Pearson code to separate functionality
